Fix <number-input> behaviour
The <number-input> returned null or a number, which made it unclear
whether the input was empty or that its contents were invalid if it
returned null. This changes it to emit a `Either<Error, Maybe<number>>`,
which allows us to:

1. Allow empty inputs in some places and disallow them in other places.
   A new `required` prop has been added to the <number-input> which, if
   set, causes the component to always emit an error if the input is
   empty. This means that inputs with the required prop set will never
   emit a `Right(Nothing)`.
2. Have standardized error messages wherever we use the <number-input>.

These changes required a lot of refactoring in every module using the
<number-input> which, at the time of writing, aren't a lot, thankfully.

This also changes the values emitted by the <assignment-submit-types>
and <submission-limits> components. They used to always emit a dict with
the selected values, but now they emit a `Either<Error, Dict>` where
`Dict` is the respective dictionary the previously emitted.