Fix NumberInput resetting its internal input to eagerly
The NumberInput always reset its internal input string when its value
prop changed, but this wasn't always correct. For example, when a user
typed "3." in the input field, it was parsed as the number "3"
(correctly), which was then emitted as a new value, but because that
would change the value prop again, it would reset its input string to
the stringified new value, in this case "3" because the toString method
on a number does not add the decimal point back.

With this change we first check if the actual numeric value has changed
before resetting the input string.

This also makes a change to utils.parseOrKeepFloat, as it would accept
the string "5a", which would return the number "5". The fix is to parse
a number with both parseFloat() and Number(), and only return a number
when both outputs are the same. This is necessary because both functions
accept distinct inputs that should not be parsed as a number but are
parsed as such anyway which is not solvable by TypeScript's type system.