DEV-103 Add exam mode (#1415)
* Make it possible to scope access tokens for a single course

* Make it possible to schedule the open date of an assignment

* Make it possible to create non LTI assignments in LTI courses

* Add better validation system for database attributes

* Add skipped state to TaskResultState

* Remove the `:max` attribute from the exam duration input

Without it it was impossible to set the input to a value greater than
the max exam duration, and users would never get to see the message that
explains why there is a limit on the duration.

* Fix submit button being enabled when it shouldn't

This happened because of two reasons:

1. We compared the deadline on the component to the deadline on the
   assignment to determine if it had changed, but in exam mode we never
   update the deadline on the component, so that would get out of sync.
2. We compared the `availableAt` date to the deadline on the component
   when calculating the new exam duration when the `availableAt` had
   changed. Now we compare to the deadline on the assignment to get
   a correct value.

* Add a (simple) changelog entry

* Add a proper Grader model

* Fix <number-input> behaviour

The <number-input> returned null or a number, which made it unclear
whether the input was empty or that its contents were invalid if it
returned null. This changes it to emit a `Either<Error, Maybe<number>>`,
which allows us to:

1. Allow empty inputs in some places and disallow them in other places.
   A new `required` prop has been added to the <number-input> which, if
   set, causes the component to always emit an error if the input is
   empty. This means that inputs with the required prop set will never
   emit a `Right(Nothing)`.
2. Have standardized error messages wherever we use the <number-input>.

These changes required a lot of refactoring in every module using the
<number-input> which, at the time of writing, aren't a lot, thankfully.

This also changes the values emitted by the <assignment-submit-types>
and <submission-limits> components. They used to always emit a dict with
the selected values, but now they emit a `Either<Error, Dict>` where
`Dict` is the respective dictionary the previously emitted.

* Add our own cached_property class

This class guarantees that the cache is cleared at the end of a request,
something that ``werkzeug.utils.cached_property`` does not guarantees.

* Improve coverage for feedback

* Fix NumberInput resetting its internal input to eagerly

The NumberInput always reset its internal input string when its value
prop changed, but this wasn't always correct. For example, when a user
typed "3." in the input field, it was parsed as the number "3"
(correctly), which was then emitted as a new value, but because that
would change the value prop again, it would reset its input string to
the stringified new value, in this case "3" because the toString method
on a number does not add the decimal point back.

With this change we first check if the actual numeric value has changed
before resetting the input string.

This also makes a change to utils.parseOrKeepFloat, as it would accept
the string "5a", which would return the number "5". The fix is to parse
a number with both parseFloat() and Number(), and only return a number
when both outputs are the same. This is necessary because both functions
accept distinct inputs that should not be parsed as a number but are
parsed as such anyway which is not solvable by TypeScript's type system.

Co-authored-by: Thomas Schaper <thoams@codegra.de>
Co-authored-by: Olmo Kramer <olmo@codegrade.com>