Fix potential diverging state between the broker and the backend
The fix is quite easy, the runner now confirms that it will start running. This
adds an extra state to the runner ("assigned"), in which we consider the runner
connected to the job, but not actually running. So we will still kill it, and
the job can still take on more runners.