Remove trailing slashes for canvasBaseUrl
This caused the URLs that were given to the backend to look like
"https://$instance.instructure.com//$data", which is not accepted by Canvas,
especially when used in things like the `aud` field in JWT tokens.

This PR also changes the style of the inputs to better match the inputs on the
newly designed pages. I needed to touch them anyway so I thought why not update
them. This is not required for this change, so definitely revert that if you
don't agree.