Make whitespace in code selectable
Because of the `user-select: none` on whitespace in the code viewer and
because the whitespace symbols were rendered on top of the actual
whitespace in the code, it was impossible to select (part of) the
whitespace. Simply ignoring all pointer events on the "fake" whitespace
that we draw on top, it is now possible to select whitespace again.