Fix getting the Brightspace LTI provider in the frontend
In the backend we used the tag "BrightSpace" for the LMS while in the
frontend we expected it to be "Brightspace" (no capital S), so we
couldn't find it and it would throw an error.

This adds a new "tag" attribute to LTI1p1 providers to map tags to the
correct LTIProvider instance.