Make sure test students are never used as peer feedback
If we had an assignment that already had a submission by the test student before
peer feedback was enabled (or if the settings of the PF was adjusted after there
was a submission by the test student) we didn't filter this user out when doing
the assignments. This PR fixes that behavior, and it makes sure we don't count
submissions by the test student when determining if we can do the initial
division.

This change also makes use of the same way to actually retrieve the users with
submissions. This method is incorrect for group assignments (which are not
support in combination with peer feedback), but is quicker as it does less
logic.