Show start date instead of deadline for exams in the header
We used to always show the deadline of an assignment in the LocalHeader,
but in the sidebar we display the available at date for exams, if the
exam hasn't started yet. This changes that so that we now always show
the start date for exams that haven't started yet.

This also gets rid of some code duplication the header on the assignment
management page.

DEV-272 #done