Disable reversing of data in plagiarism cases (#1468)
This is no longer now that we guarantee in the backend that the case that was
submitted to the assignment on which the run was done is the first in each of
the data arrays. This caused bugs that caused the wrong files to be displayed
for the user.

This also adds a not `null` contraint on the `plagiarism_case_id` column, which
never should be `null` anyway (and never is in any production instance).