DEV-275 Sort permissions by display name (#1464)
* Sort permissions by display name

The permissions list was sorted by the internal name of permissions, e.g.
`can_edit_others_work`, but because we have made mistakes in the past in naming
permissions inconsistently, such as naming a permission `manage_rubrics`
instead of `can_manage_rubrics`, which caused it to be sorted at the end of the
list. But because we never include `Can` in the display name of a permission,
the items in the list weren't ordered alphabetically.

* Ignore unknown permissions in the frontend instead of crashing

* Improve searching permissions

We now use `src/utils/search.ts` when searching permissions, so you can now
search by their name, their description, or their warning message, if they have
one.

You can also filter permissions by their state for a specific role. For
example, searching for "Student:true" will list only the permissions that are
enabled for the "Student" role. "Teacher:true Student:false" will list all
permissions that are enabled for the "Teacher" but disabled for the "Student".

* Search by substring instead of regex in utils/search.ts

Searching with regexes was prone to errors because we had to escape the input
string to be a valid regex, which has already caused some errors.

DEV-275 #done

Co-authored-by: Thomas Schaper <thomas@codegra.de>