Fix issue where runners in the ``assigned`` state were not killed
This happened because of a race condition. If we have two runners both looking
to run a job, and only one job (that wants one runner). It could happen that
both runners would get assigned to the single job in the broker, but that one
request to the application backend failed. In this case the broker would keep
suggesting the assigned job, which the application backend would never consider
as an option it already has enough runners.