Fix parsing roles in Open edX
Students only have the role "Student" in Open edX (which isn't even a valid
course role in LTI 1.1), and some teachers have the bare "Administrator"
role. This adds a translation for the "Student" role and starts parsing the bare
"Administrator" role for Open edX, it doesn't change anything for other LMSes.