Correctly keep track of rubric rows with tracking ids
When updating a rubric, we would simply overwrite the previous rubric in
the store with the new data received from the server. This is, however,
problematic, because the tracking ids of the rubric rows weren't kept in
sync, causing glitches in the <transition> of the new rubric editor.

With this change, instead of overwriting the previous rubric in the
store, we update it with the new data we have received from the server,
so the tracking ids stay the same.