Fix the CodeViewer when loading settings fails (#1496)
In Google Chrome in incognito mode the language store would return `undefined`
when getting the language, which we would then store in `selectedLanguage`,
but in the rest of the component we check with `this.selectedLanguage != null`
to make sure we already loaded the language, and of course `null ==
undefined`. This PR fixes that behavior by also checking the return from
`getItem` against `null` with eqeq and not eqeqeq. Furthermore, this also adds
a catch clause on the call to `getItem` to make sure we never crash/freeze the
page because of that function rejecting.